Move thread pool to qurt_thread_pool.cpp.
