+++ src/runtime/hexagon_remote/Makefile
@@ -98,20 +98,24 @@ bin/%/halide_hexagon_remote_skel.o: bin/src/halide_hexagon_remote_skel.c
 
 bin/%/halide_remote.o: halide_remote.cpp
 	mkdir -p $(@D)
-	$(CXX-$*) $(CCFLAGS-$*) -fPIC -c halide_remote.cpp -o $@
+	$(CXX-$*) $(CCFLAGS-$*) -fPIC -c $^ -o $@
+
+bin/%/qurt_thread_pool.o: qurt_thread_pool.cpp
+	mkdir -p $(@D)
+	$(CXX-$*) $(CCFLAGS-$*) -fPIC -c $^ -o $@
 
 bin/%/halide_host.o: halide_host.cpp
 	mkdir -p $(@D)
-	$(CXX-$*) $(CCFLAGS-$*) -fPIC -c halide_host.cpp -o $@
+	$(CXX-$*) $(CCFLAGS-$*) -fPIC -c $^ -o $@
 
 # Build rules for the hexagon implementation.
-bin/%/libhalide_hexagon_remote_skel.so: bin/%/halide_remote.o bin/%/halide_hexagon_remote_skel.o
+bin/%/libhalide_hexagon_remote_skel.so: bin/%/halide_remote.o bin/%/qurt_thread_pool.o bin/%/halide_hexagon_remote_skel.o
 	mkdir -p $(@D)
 	$(CC-$*) -m$* -O0 -mG0lib -G0 -fpic -shared -lc -o $@ \
 	    -Wl,-Bsymbolic -Wl,--wrap=malloc -Wl,--wrap=calloc -Wl,--wrap=free \
 	    -Wl,--wrap=realloc -Wl,--wrap=memalign -Wl,--wrap=__stack_chk_fail \
 	    -Wl,-soname=libhalide_hexagon_remote_skel.so \
-	    -Wl,--start-group  bin/$*/halide_remote.o bin/$*/halide_hexagon_remote_skel.o -Wl,--end-group
+	    -Wl,--start-group  bin/$*/halide_remote.o bin/$*/qurt_thread_pool.o bin/$*/halide_hexagon_remote_skel.o -Wl,--end-group
 
 bin/%/libhalide_hexagon_host.so: bin/src/halide_hexagon_remote_stub.c bin/%/halide_host.o
 	mkdir -p $(@D)
