+++ src/runtime/hexagon_remote/halide_remote.cpp
@@ -21,63 +21,6 @@ extern "C" {
 typedef halide_hexagon_remote_handle_t handle_t;
 typedef halide_hexagon_remote_buffer buffer;
 
-#define MAX_WORKER_THREADS 4
-#define NUM_WORKER_THREADS_TO_CREATE (MAX_WORKER_THREADS - 1)
-#define STACK_SIZE 256 * 1024
-
-static char stack [MAX_WORKER_THREADS][STACK_SIZE] __attribute__ ((__aligned__(128)));
-qurt_sem_t wait_for_work;
-//qurt_cond_t work_in_queue;
-struct work {
-    work *next_job;
-    int (*f)(void *, int, uint8_t *);
-    void *user_context;
-    int next, end;
-    uint8_t *closure;
-    int active_workers;
-    int exit_status;
-    qurt_cond_t wakeup_owner;
-    qurt_hvx_mode_t curr_hvx_mode;
-    // A job can be in the following states.
-    // claimed - When the thread_pool has started work on the entire job, but not necessarily
-    //           completed it. Condition: next >= end.
-    // running - When there are active workers on the job. Condition: active_workers > 0;
-    // done - When the job is completely done and there are no active workers on the job.
-    bool claimed() { return next >= end; }
-    bool running() { return active_workers > 0; }
-    bool done() { return claimed() && !running(); }
-};
-bool thread_pool_initialized = false;
-struct work_queue_t {
-    // all fields are protected by this mutex.
-    qurt_mutex_t work_mutex;
-
-    // Jobs that the thread pool needs to work on.
-    work *jobs;
-
-    // Global flag indicating if the thread pool has been shutdown.
-    bool shutdown;
-
-    bool running() {
-        return !shutdown;
-    }
-    void lock() {
-        qurt_mutex_lock(&work_mutex);
-    }
-    void unlock() {
-        qurt_mutex_unlock(&work_mutex);
-    }
-    void init() {
-        qurt_mutex_init(&work_mutex);
-        lock();
-        shutdown = false;
-        jobs = NULL;
-        unlock();
-    }
-};
-work_queue_t work_queue;
-qurt_thread_t threads[MAX_WORKER_THREADS];
-
 extern "C" {
 
 void halide_print(void *user_context, const char *str) {
@@ -100,193 +43,6 @@ void halide_free(void *user_context, void *ptr) {
 int halide_do_task(void *user_context, halide_task_t f, int idx,
                    uint8_t *closure) {
     return f(user_context, idx, closure);
-}
-
-// This function does the real work of the thread pool.
-// owned_job is used to tell the differece between the master thread
-// and the worker threads. If owned_job is non-null then it means that
-// we are in the master thread, i.e. the thread that put owned_job in
-// the work queue. This thread, like other worker threads in this function,
-// tries to acquire a lock on the work queue. Each thread then looks for
-// a job to do. In the case of the master thread, it is "owned_job" while
-// a worker thread has to iterate backwards over the work queue and find
-// a job to work on. Once a job has been found, the thread, after some
-// book-keeping, releases the lock and call halide_do_task to do the job.
-// After the job is done, the lock on the queue is re-acquired to update
-// the status and the thread loops again to look for new work. If a worker
-// thread doesn't find any new work, it goes to sleep until awoken by
-// the master thread.
-void goto_work(work *owned_job) {
-    // You can work only if you get a lock on the work queue.
-    unsigned int tid = qurt_thread_get_id();
-
-    bool locked = owned_job != NULL;
-
-    // ***********************
-    // *** Lock work queue ***
-    // ***********************
-    work_queue.lock();
-
-    // If I'm a job owner, then I was the thread that called
-    // do_par_for, and I should only stay in this function while there
-    // are active workers on the job (i.e. running()) or if the job
-    // hasn't been claimed entirely (!claimed()). If I'm a lowly worker
-    // thread, I should stay in this function as long as the work queue is running.
-    while (owned_job != NULL ? (owned_job->running() || !owned_job->claimed())
-           : work_queue.running()) {
-
-        work *job = NULL;
-        if (!owned_job) {
-            // If this threads doesn't own a job, it looks for one
-            // and tries to do it. If it cannot find a job, it goes
-            // to sleep.
-            job = work_queue.jobs;
-            if (!job) {
-                // release the lock and go to sleep.
-                // ***************************
-                // *** Work queue unlocked ***
-                // ***************************
-                work_queue.unlock();
-                qurt_sem_down(&wait_for_work);
-                 // ***********************
-                // *** Lock work queue ***
-                // ***********************
-                work_queue.lock();
-                continue;
-            }
-        } else {
-            // We are here only if this thread owns a job that is not done.
-            // So do a part of it.
-            if (owned_job->claimed()) {
-                // the owner goes to sleep till the workers wake it up.
-                qurt_cond_wait(&owned_job->wakeup_owner, &work_queue.work_mutex);
-                // This thread should have the lock now after having been woken up.
-                break;
-            } else {
-                job = owned_job;
-            }
-        }
-        int myjob = job->next++;
-        job->active_workers++;
-        // If all tasks of the job have been claimed, then pop the job off the stack.
-        if (job->claimed()) {
-            work_queue.jobs = job->next_job;
-        }
-        // ***************************
-        // *** Work queue unlocked ***
-        // ***************************
-        work_queue.unlock();
-
-        if (!locked) {
-            int lock_status = qurt_hvx_lock(job->curr_hvx_mode);
-            if (lock_status != QURT_EOK) {
-                FARF(LOW, "HVX_TP: %d: goto_work: qurt_hvx_lock(%d) failed\n", tid, job->curr_hvx_mode);
-            } else {
-                FARF(LOW, "HVX_TP: %d: goto_work: qurt_hvx_lock(%d) succeeded.\n", tid, job->curr_hvx_mode);
-            }
-        }
-
-        FARF(LOW, "HVX_TP: %d: goto_work: About to do_task, user_context = 0x%x, f = 0x%x x = %d \n", tid, job->user_context, job->f, myjob);
-        int result = halide_do_task(job->user_context, job->f,
-                                    myjob, job->closure);
-        FARF(LOW, "HVX_TP: %d: goto_work: Finished do_task with status = %d\n", tid, result);
-
-        qurt_hvx_unlock();
-        locked = false;
-
-        // ***********************
-        // *** Lock work queue ***
-        // ***********************
-        work_queue.lock();
-        if (result) {
-            job->exit_status = result;
-            job->next = job->end;
-            work_queue.jobs = job->next_job;
-            FARF(LOW, "HVX_TP: %d: goto_work: halide_do_task returned nonzero result = %d\n", tid, result);
-        }
-        job->active_workers--;
-        if (job->done()) {
-            if (!owned_job) {
-                qurt_cond_signal(&(job->wakeup_owner));
-            }
-        }
-    }
-    // ***************************
-    // *** Work queue unlocked ***
-    // ***************************
-    work_queue.unlock();
-}
-void thread_server(void *arg) {
-    goto_work(NULL);
-    qurt_thread_exit(QURT_EOK);
-}
-void create_threads(int num_threads) {
-    // Acquire a lock on the work queue.
-    qurt_thread_attr_t thread_attr;
-    for (int i = 0; i < num_threads; ++i) {
-        qurt_thread_attr_init(&thread_attr);
-        qurt_thread_attr_set_stack_addr(&thread_attr, stack[i]);
-        qurt_thread_attr_set_stack_size(&thread_attr, STACK_SIZE);
-        qurt_thread_attr_set_priority(&thread_attr, 100);
-        qurt_thread_create(&threads[i], &thread_attr, thread_server, (void *)i);
-    }
-}
-void qurt_thread_pool_init() {
-    qurt_sem_init_val(&wait_for_work, 0);
-    work_queue.init();
-    create_threads(NUM_WORKER_THREADS_TO_CREATE);
-    thread_pool_initialized = true;
-}
-
-int halide_do_par_for(void *user_context, halide_task_t f,
-                      int min, int size, uint8_t *closure) {
-
-    // 1. If the thread pool hasn't been initialized, initiliaze it.
-    // This involves.
-    //    a) Creating the threads.
-    //    b) Acquiring a lock on the work queue and clearing the jobs therein
-    //    c) Setting up a semaphore on which worker threads sleep until awoken
-    //       by this thread i.e. the master thread.
-    if (!thread_pool_initialized) {
-        // Initialize the work queue mutex.
-        // Initialize the wait_for_work semaphore.
-        // lock work queue
-        //    wq.shutdown = false;
-        //    wq.jobs = NULL;
-        // unlcok work queue
-        // create NUM_WORKER_THREADS_TO_CREATE number of threads.
-        // thread_pool_initialized = true;
-        qurt_thread_pool_init();
-    }
-
-    // 2. Lock the work queue again.
-    work_queue.lock();
-
-    // 3. Put work in the global work queue.
-    work job;
-    job.f = f;
-    job.user_context = user_context;
-    job.next = min;          // Start at this index.
-    job.end  = min + size;   // Keep going until one less than this index.
-    job.closure = closure;   // Use this closure.
-    job.exit_status = 0;     // The job hasn't failed yet
-    job.active_workers = 0;  // Nobody is working on this yet
-    qurt_cond_init(&job.wakeup_owner);
-    job.next_job = work_queue.jobs;
-    job.curr_hvx_mode = (qurt_hvx_mode_t) qurt_hvx_get_mode();
-    work_queue.jobs = &job;
-
-    // 4. Unlock global work queue.
-    work_queue.unlock();
-
-    // 5. Wake up the other threads in the pool.
-    qurt_sem_add(&wait_for_work, size);
-
-    // 6. Do some work in the master queue.
-    goto_work(&job);
-
-    qurt_hvx_lock(job.curr_hvx_mode);
-    return job.exit_status;
 }
 
 void *halide_get_symbol(const char *name) {
@@ -444,18 +200,8 @@ int halide_hexagon_remote_run(handle_t module_ptr, handle_t function,
 int halide_hexagon_remote_release_kernels(handle_t module_ptr, int codeLen) {
     dlclose(reinterpret_cast<void*>(module_ptr));
     if (context_count-- == 0) {
-        if (thread_pool_initialized) {
-            work_queue.lock();
-            work_queue.jobs = NULL;
-            work_queue.shutdown = true;
-            qurt_sem_add(&wait_for_work, NUM_WORKER_THREADS_TO_CREATE);
-            work_queue.unlock();
-            thread_pool_initialized = false;
-            for (int i = 0; i < NUM_WORKER_THREADS_TO_CREATE; ++i) {
-                int status;
-                qurt_thread_join(threads[i], &status);
-            }
-        }
+        halide_shutdown_thread_pool();
+
         HAP_power_request(0, 0, -1);
     }
     return 0;
